LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-from-literal-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: ReferenceError: Can't find variable: TrustedHTML

Harness Error (FAIL), message = ReferenceError: Can't find variable: TrustedHTML
